cleanup

+++ pkg/gitops/types/gitops_interface.go
@@ -1,5 +1,5 @@
 package types
 
 type DownstreamGitOps interface {
-	CreateGitOpsDownstreamCommit(appID string, clusterID string, newSequence int, archiveDir string, downstreamName string) (string, error)
+	CreateGitOpsDownstreamCommit(appID string, clusterID string, newSequence int64, archiveDir string, downstreamName string) (string, error)
 }
